[tests] decouple FixtureSplitter
