+++ packages/DeadCode/src/Rector/Assign/RemoveDoubleAssignRector.php
@@ -2,8 +2,6 @@
 
 namespace Rector\DeadCode\Rector\Assign;
 
-use PhpParser\Node\Stmt\Catch_;
-use PhpParser\Node\Stmt\Case_;
 use PhpParser\Node;
 use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\FuncCall;
@@ -11,6 +9,8 @@ use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\PropertyFetch;
 use PhpParser\Node\Expr\StaticCall;
 use PhpParser\Node\Expr\Variable;
+use PhpParser\Node\Stmt\Case_;
+use PhpParser\Node\Stmt\Catch_;
 use PhpParser\Node\Stmt\Do_;
 use PhpParser\Node\Stmt\Else_;
 use PhpParser\Node\Stmt\ElseIf_;
