+++ tests/Rector/Argument/ArgumentAdderRector/Fixture/already_added.php.inc
@@ -2,9 +2,9 @@
 
 namespace Rector\Tests\Rector\Argument\ArgumentAdderRector\Fixture;
 
-use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\ContainerBuilder;
+use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\SomeContainerBuilder;
 
-class AlreadyAdded extends ContainerBuilder
+class AlreadyAdded extends SomeContainerBuilder
 {
     public function addCompilerPass($pass, $type, int $priority = 10)
     {
