+++ tests/Rector/Argument/ArgumentAdderRector/Fixture/fixture.php.inc
@@ -2,13 +2,13 @@
 
 namespace Rector\Tests\Rector\Argument\ArgumentAdderRector\Fixture;
 
-use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\ContainerBuilder;
+use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\SomeContainerBuilder;
 
 class SomeClass
 {
     public function create()
     {
-        $containerBuilder = new ContainerBuilder();
+        $containerBuilder = new SomeContainerBuilder();
         $containerBuilder->compile();
     }
 }
@@ -19,13 +19,13 @@ class SomeClass
 
 namespace Rector\Tests\Rector\Argument\ArgumentAdderRector\Fixture;
 
-use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\ContainerBuilder;
+use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\SomeContainerBuilder;
 
 class SomeClass
 {
     public function create()
     {
-        $containerBuilder = new ContainerBuilder();
+        $containerBuilder = new SomeContainerBuilder();
         $containerBuilder->compile(false);
     }
 }
