+++ tests/Rector/Argument/ArgumentAdderRector/Fixture/fixture3.php.inc
@@ -2,9 +2,9 @@
 
 namespace Rector\Tests\Rector\Argument\ArgumentAdderRector\Fixture;
 
-use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\ContainerBuilder;
+use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\SomeContainerBuilder;
 
-class CustomContainerBuilder extends ContainerBuilder
+class CustomContainerBuilder extends SomeContainerBuilder
 {
     public function addCompilerPass($pass, $type)
     {
@@ -17,9 +17,9 @@ class CustomContainerBuilder extends ContainerBuilder
 
 namespace Rector\Tests\Rector\Argument\ArgumentAdderRector\Fixture;
 
-use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\ContainerBuilder;
+use Rector\Tests\Rector\Argument\ArgumentAdderRector\Source\SomeContainerBuilder;
 
-class CustomContainerBuilder extends ContainerBuilder
+class CustomContainerBuilder extends SomeContainerBuilder
 {
     public function addCompilerPass($pass, $type, int $priority = 0)
     {
